Removing "Location" and "Data" tabs from CSV
We don't have anything in the Location data field at the moment, so I'm removing it for the present purposes of having something presentable for the workshop next week. I'm curious to have a conversation about how we can use the distinction that CB provides for us between "Location" and "Map" to potentially enliven or facilitate our map-related deliberations.

I feel pretty strongly that we should not have an option for an external user to download all of our data in the single click of a button. This is a bigger conversation, too. The short version from my current perspective is that it should all be open access and available to people who want to talk to us about it, or I guess at this stage to download it from GitHub, but not for the sort of reckless consumption of datapoints. This raises a question about GitHub and how we have things here in CSV rather than JSON(?) I simply do not know enough about this but am thinking about (Un)silencing Slavery.